cq01_nnhai_ass4/requirements.txt
django==3.2.18
djangorestframework==3.14.0
djongo==1.3.6
pymongo==3.12.3
mysqlclient==2.1.0
psycopg2==2.9.3
python-dotenv==1.0.0
requests==2.31.0
django-cors-headers==4.1.0
Pillow==10.0.0
django-crispy-forms==2.0
django-allauth==0.54.0
django-rest-auth==0.9.5
django-filter==23.2
django-storages==1.14
boto3==1.28.0
celery==5.3.1
redis==4.6.0
gunicorn==21.2.0
whitenoise==6.5.0 